kernel: fix io_pgetevents OOB read, COW perf counter, and all warnings
io_pgetevents read args->arg[6] past the 6-element argument array on every
call: on 64-bit ABIs Linux drops the sigsetsize argument entirely, so pass
0 for the (void)ed parameter instead.

handle_cow_fault incremented the global page-fault counter unconditionally
while the per-task counter only counted successes (and the equivalent
file-fault path counts both inside the success branch); brace both.

Warning cleanup across the tree (riscv64 and loongarch64 BRINGUP builds now
compile with zero warnings, then -Werror is enabled for the kernel by
default with a KERNEL_WERROR=0 escape hatch for unverified arches):
- unused params/vars/functions: sched_runq_steal_locked lrq, exec
  native_start_info (now NATIVE-only), proc_debug_singlestep sig,
  proc_debug_getregset kind, posix_mq now, kep imm, udisk pid, xattr
  nofollow, landlock flags, drm ioctl ctx/arg, dead uart
  uart_task_is_interrupt_target, dead cmp_u64 and pcache_hash_remove
- sign/const: sys_ptrace regs count comparison, ramfs overlay content
  cast, udriver window loops (int + (int) macro so NR=0 configs warn
  neither way), drm magic via uintptr_t, exec vdso_map_image const cast
- nested comment markers in psi.h/alsa.h, intentional
  __builtin_return_address(2) suppressed locally, intentional inotify
  alias override suppressed locally

+++ kernel/abi/linux/sys_ptrace.c
@@ -79,7 +79,7 @@ int64_t sys_ptrace(int request, int pid, void *addr, void *data)
             return -EIO;
         unsigned long off = (unsigned long)addr;
         if (off % sizeof(unsigned long) ||
-            off / sizeof(unsigned long) >= linux_arch_regs_count())
+            off / sizeof(unsigned long) >= (unsigned long)linux_arch_regs_count())
             return -EIO;
         proc_debug_regs_t regs;
         int ret = proc_debug_getregs(pid, &regs);
@@ -95,7 +95,7 @@ int64_t sys_ptrace(int request, int pid, void *addr, void *data)
             return -EIO;
         unsigned long off = (unsigned long)addr;
         if (off % sizeof(unsigned long) ||
-            off / sizeof(unsigned long) >= linux_arch_regs_count())
+            off / sizeof(unsigned long) >= (unsigned long)linux_arch_regs_count())
             return -EIO;
         proc_debug_regs_t regs;
         int ret = proc_debug_getregs(pid, &regs);
